+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generate Python client
 - Generate Ruby client
 - Add Gemfile
+- Reduce operation id to 70 chars